# docs/swagger.yaml
definitions:
  handler.request:
    properties:
      codigo:
        type: string
      cor:
        type: string
      data_de_criacao:
        type: string
      estoque:
        type: integer
      nome:
        type: string
      preco:
        type: integer
      publicacao:
        type: boolean
    type: object
  web.Response:
    properties:
      code:
        type: string
      data: {}
      error:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: API Support
    url: https://developers.mercadolibre.com.ar/support
  description: This API Handle MELI Products.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://developers.mercadolibre.com.ar/es_ar/terminos-y-condiciones
  title: MELI Bootcamp API
  version: "1.0"
paths:
  /products:
    get:
      consumes:
      - application/json
      description: get products
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
      summary: List products
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: insert products
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: Product to insert
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/handler.request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
      summary: Insert products
      tags:
      - Products
  /products/{id}:
    delete:
      consumes:
      - application/json
      description: delete a product
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/web.Response'
      summary: Delete a product
      tags:
      - Products
    patch:
      consumes:
      - application/json
      description: update product's name and price
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product to update
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/handler.request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
      summary: Update a product's name and price
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: update product details
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product to update
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/handler.request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
      summary: Update a product
      tags:
      - Products
swagger: "2.0"